factor out storybook layout plugin
