+++ packages/plugins/plugin-markdown/package.json
@@ -67,6 +67,7 @@
   },
   "devDependencies": {
     "@dxos/debug": "workspace:*",
+    "@dxos/plugin-storybook-layout": "workspace:*",
     "@dxos/plugin-theme": "workspace:*",
     "@dxos/random": "workspace:*",
     "@dxos/react-ui": "workspace:*",
