+++ packages/plugins/plugin-markdown/tsconfig.json
@@ -109,6 +109,9 @@
       "path": "../plugin-space"
     },
     {
+      "path": "../plugin-storybook-layout"
+    },
+    {
       "path": "../plugin-theme"
     }
   ]
