+++ packages/plugins/plugin-storybook-layout/README.md
@@ -1,0 +1,3 @@
+# @dxos/plugin-storybook-layout
+
+A plugin that provides a simple layout to integrate and test other plugins in storybook without bringing in the complexity of @dxos/app-graph.
