+++ packages/plugins/plugin-storybook-layout/src/capabilities/index.ts
@@ -1,0 +1,10 @@
+//
+// Copyright 2025 DXOS.org
+//
+
+import { lazy } from '@dxos/app-framework';
+
+export const IntentResolver = lazy(() => import('./intent-resolver'));
+export const State = lazy(() => import('./state'));
+
+export { LayoutState } from './state';
