+++ packages/plugins/plugin-storybook-layout/src/index.ts
@@ -1,0 +1,6 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+export * from './meta';
+export * from './StorybookLayoutPlugin';
