+++ packages/plugins/plugin-storybook-layout/src/meta.ts
@@ -1,0 +1,11 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+import { type PluginMeta } from '@dxos/app-framework';
+
+export const meta: PluginMeta = {
+  id: 'dxos.org/plugin/storybook-layout',
+  name: 'Storybook Layout',
+  source: 'https://github.com/dxos/dxos/tree/main/packages/plugins/plugin-storybook-layout',
+};
