+++ release-please-config.json
@@ -639,6 +639,11 @@
         },
         {
           "type": "json",
+          "path": "packages/plugins/plugin-storybook-layout/package.json",
+          "jsonpath": "$.version"
+        },
+        {
+          "type": "json",
           "path": "packages/plugins/plugin-table/package.json",
           "jsonpath": "$.version"
         },
