+++ tsconfig.paths.json
@@ -211,6 +211,9 @@
       "@dxos/plugin-status-bar": [
         "packages/plugins/plugin-status-bar/src/index.ts"
       ],
+      "@dxos/plugin-storybook-layout": [
+        "packages/plugins/plugin-storybook-layout/src/index.ts"
+      ],
       "@dxos/plugin-table": [
         "packages/plugins/plugin-table/src/index.ts"
       ],
